Remove mismatch type method longValue() -> should be Double

+++ src/main/java/com/example/Event_Ticket_System/service/BookingService.java
@@ -95,7 +95,7 @@ public class BookingService {
                 ? booking.getTicketType().getEvent().getTitle() : null);
         dto.setTicket_type_name(booking.getTicketType() != null ? booking.getTicketType().getName() : null);
         dto.setPrice(booking.getTicketType() != null && booking.getTicketType().getPrice() != null
-                ? booking.getTicketType().getPrice().longValue() : null);
+                ? booking.getTicketType().getPrice() : null);
         return dto;
     }
 }
